created a useForm custom hook that handles the stateful logic from CheckoutForm. Imported useForm in CheckoutForm, and handled the logic by the useForm custom hook.
